requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
joblib
requests
beautifulsoup4
tldextract
python-dotenv
